requirements.txt
openai
sentence-transformers
spacy
opencv-python
yt-dlp
openai-whisper
inflect
psutil